Removing broken links
* Removed all links to the trading questions page, and other pages which are not present in this repo
* Removed the broken link callouts at the top of each page
* Changed "What's GitPod?" link to be relative, which should be fine as it's currently only used in the How-to area

If a whole sentence was leading up to a link (e.g. "See the Trading questions section for more detail on order types") then I removed the whole sentence. If the link was _in_ a sentence that makes sense without the link, I left the sentence intact, only removing the link itself.