Update plugin and docusaurus
The version of the openapi-plugin that was installed by default on this
branch doesn't support the 'hide_send_button' frontmatter that can hide
the 'Send API request' button on a per-page basis. I couldn't get a
working version of the updated plugin working without just upgrading to
the latest version. So the API pages look a little different (style, not
layout) and have a few more features (e.g. headers in requests).

- Update to docusaurus 2.4.0
- Update openapi plugin to 2.0.0-beta2